# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Kryv live streaming platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

/**
 * live_game categories power Kryv Live's browse sidebar; genre categories power Kryv Watch/Cinema
 */
export type CategoryKind = typeof CategoryKind[keyof typeof CategoryKind];


export const CategoryKind = {
  live_game: 'live_game',
  genre: 'genre',
} as const;

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** live_game categories power Kryv Live's browse sidebar; genre categories power Kryv Watch/Cinema */
  kind: CategoryKind;
  /** @nullable */
  imageUrl: string | null;
  liveChannelCount: number;
  viewerCount: number;
}

export interface ChannelSummary {
  id: number;
  slug: string;
  displayName: string;
  /** @nullable */
  avatarUrl: string | null;
  /** @nullable */
  bannerUrl: string | null;
  /** @nullable */
  streamTitle: string | null;
  isLive: boolean;
  viewerCount: number;
  followerCount: number;
  /** @nullable */
  categoryId: number | null;
  /** @nullable */
  categoryName: string | null;
  /**
     * Mux playback id used to build the HLS playback URL. Null until the channel has gone live at least once.
     * @nullable
     */
  playbackId: string | null;
}

export type ChannelDetail = ChannelSummary & ({
  /** @nullable */
  description: string | null;
  isFollowing: boolean;
  isOwner: boolean;
  ownerUserId: string;
  createdAt: string;
});

export interface ChannelInput {
  /**
     * @minLength 1
     * @maxLength 60
     */
  displayName: string;
  /** @maxLength 500 */
  description?: string;
  categoryId?: number;
}

export interface ChannelUpdate {
  /**
     * @minLength 1
     * @maxLength 60
     */
  displayName?: string;
  /** @maxLength 500 */
  description?: string;
  avatarUrl?: string;
  bannerUrl?: string;
  categoryId?: number;
  /** @maxLength 140 */
  streamTitle?: string;
}

export interface StreamCredentials {
  /** RTMP ingest server URL to enter in broadcasting software (e.g. OBS) */
  rtmpUrl: string;
  /** Secret stream key — combined with rtmpUrl to start broadcasting. Never shown to anyone but the owner. */
  streamKey: string;
  /** Public Mux playback id for this channel's live stream */
  playbackId: string;
}

export type VideoSummaryContentType = typeof VideoSummaryContentType[keyof typeof VideoSummaryContentType];


export const VideoSummaryContentType = {
  upload: 'upload',
  original: 'original',
} as const;

export type VideoSummaryUploadStatus = typeof VideoSummaryUploadStatus[keyof typeof VideoSummaryUploadStatus];


export const VideoSummaryUploadStatus = {
  waiting: 'waiting',
  processing: 'processing',
  ready: 'ready',
  errored: 'errored',
} as const;

export interface VideoSummary {
  id: number;
  title: string;
  /**
     * Landscape 16:9 thumbnail — used in Kryv Watch grids
     * @nullable
     */
  thumbnailUrl: string | null;
  /**
     * Portrait 2:3 poster — used in Kryv Cinema rows
     * @nullable
     */
  posterUrl: string | null;
  /**
     * Wide cinematic backdrop — used in Kryv Cinema hero banners
     * @nullable
     */
  backdropUrl: string | null;
  /** @nullable */
  durationSeconds: number | null;
  viewCount: number;
  channelId: number;
  channelName: string;
  /** @nullable */
  channelAvatarUrl: string | null;
  /** @nullable */
  categoryId: number | null;
  /** @nullable */
  categoryName: string | null;
  contentType: VideoSummaryContentType;
  /**
     * Mux playback id for HLS playback. Null until Mux finishes processing the upload.
     * @nullable
     */
  playbackId: string | null;
  uploadStatus: VideoSummaryUploadStatus;
  createdAt: string;
}

export type VideoDetail = VideoSummary & ({
  /** @nullable */
  description: string | null;
  isOwner: boolean;
});

export type VideoInputContentType = typeof VideoInputContentType[keyof typeof VideoInputContentType];


export const VideoInputContentType = {
  upload: 'upload',
  original: 'original',
} as const;

export interface VideoInput {
  /**
     * @minLength 1
     * @maxLength 100
     */
  title: string;
  /** @maxLength 5000 */
  description?: string;
  categoryId?: number;
  contentType?: VideoInputContentType;
}

export type VideoCreateResponse = VideoDetail & {
  /** Mux direct-upload URL — PUT the raw video file here from the browser */
  uploadUrl: string;
};

export interface VideoUpdate {
  /**
     * @minLength 1
     * @maxLength 100
     */
  title?: string;
  /** @maxLength 5000 */
  description?: string;
  categoryId?: number;
}

export interface CinemaRow {
  title: string;
  items: VideoSummary[];
}

export interface CinemaHome {
  hero: VideoSummary | null;
  rows: CinemaRow[];
}

export interface ChatMessage {
  id: number;
  channelId: number;
  userId: string;
  username: string;
  /** @nullable */
  avatarUrl: string | null;
  message: string;
  createdAt: string;
}

export interface ChatMessageInput {
  /**
     * @minLength 1
     * @maxLength 500
     */
  message: string;
}

export interface Me {
  id: string;
  username: string;
  /** @nullable */
  avatarUrl: string | null;
  channel: ChannelSummary | null;
  followedChannels: ChannelSummary[];
}

export interface DiscoverSummary {
  featuredChannels: ChannelSummary[];
  topCategories: Category[];
  totalLiveChannels: number;
  totalViewers: number;
}

export type ListCategoriesParams = {
kind?: ListCategoriesKind;
};

export type ListCategoriesKind = typeof ListCategoriesKind[keyof typeof ListCategoriesKind];


export const ListCategoriesKind = {
  live_game: 'live_game',
  genre: 'genre',
} as const;

export type ListChannelsParams = {
categorySlug?: string;
live?: boolean;
search?: string;
};

export type ListVideosParams = {
channelId?: number;
categorySlug?: string;
contentType?: ListVideosContentType;
search?: string;
};

export type ListVideosContentType = typeof ListVideosContentType[keyof typeof ListVideosContentType];


export const ListVideosContentType = {
  upload: 'upload',
  original: 'original',
} as const;