# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Kryv live streaming platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /channels:
    get:
      operationId: listChannels
      tags: [channels]
      summary: Browse channels, optionally filtered by category, live status, or search
      parameters:
        - name: categorySlug
          in: query
          required: false
          schema:
            type: string
        - name: live
          in: query
          required: false
          schema:
            type: boolean
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of channels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChannelSummary"
    post:
      operationId: createChannel
      tags: [channels]
      summary: Create a channel for the current user (become a creator)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChannelInput"
      responses:
        "201":
          description: Created channel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelDetail"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: User already has a channel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /channels/{id}:
    get:
      operationId: getChannel
      tags: [channels]
      summary: Get a single channel's detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Channel detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelDetail"
        "404":
          description: Channel not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateChannel
      tags: [channels]
      summary: Update the current user's own channel (profile, category, offline)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChannelUpdate"
      responses:
        "200":
          description: Updated channel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelDetail"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not the channel owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Channel not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /channels/{id}/stream:
    post:
      operationId: createChannelStream
      tags: [channels]
      summary: Create or fetch this channel's live ingest credentials (RTMP URL, stream key, playback id)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ingest credentials for broadcasting software (OBS, etc.)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamCredentials"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not the channel owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Channel not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /channels/{id}/follow:
    post:
      operationId: followChannel
      tags: [channels]
      summary: Follow a channel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated channel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelDetail"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Channel not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: unfollowChannel
      tags: [channels]
      summary: Unfollow a channel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated channel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelDetail"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Channel not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /channels/{id}/messages:
    get:
      operationId: listChannelMessages
      tags: [chat]
      summary: List recent chat messages for a channel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recent chat messages, oldest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
        "404":
          description: Channel not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: createChannelMessage
      tags: [chat]
      summary: Post a chat message to a channel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "201":
          description: Created chat message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Channel not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    ChannelSummary:
      type: object
      required:
        [
          id,
          slug,
          displayName,
          avatarUrl,
          bannerUrl,
          streamTitle,
          isLive,
          viewerCount,
          followerCount,
          categoryId,
          categoryName,
          playbackId,
        ]
      properties:
        id: { type: integer }
        slug: { type: string }
        displayName: { type: string }
        avatarUrl: { type: ["string", "null"] }
        bannerUrl: { type: ["string", "null"] }
        streamTitle: { type: ["string", "null"] }
        isLive: { type: boolean }
        viewerCount: { type: integer }
        followerCount: { type: integer }
        categoryId: { type: ["integer", "null"] }
        categoryName: { type: ["string", "null"] }
        playbackId:
          description: Mux playback id used to build the HLS playback URL. Null until the channel has gone live at least once.
          type: ["string", "null"]
    ChannelDetail:
      allOf:
        - $ref: "#/components/schemas/ChannelSummary"
        - type: object
          required: [description, isFollowing, isOwner, ownerUserId, createdAt]
          properties:
            description: { type: ["string", "null"] }
            isFollowing: { type: boolean }
            isOwner: { type: boolean }
            ownerUserId: { type: string }
            createdAt: { type: string, format: date-time }
    ChannelInput:
      type: object
      required: [displayName]
      properties:
        displayName: { type: string, minLength: 1, maxLength: 60 }
        description: { type: string, maxLength: 500 }
        categoryId: { type: integer }
    ChannelUpdate:
      type: object
      properties:
        displayName: { type: string, minLength: 1, maxLength: 60 }
        description: { type: string, maxLength: 500 }
        avatarUrl: { type: string }
        bannerUrl: { type: string }
        categoryId: { type: integer }
        streamTitle: { type: string, maxLength: 140 }
    StreamCredentials:
      type: object
      required: [rtmpUrl, streamKey, playbackId]
      properties:
        rtmpUrl:
          description: RTMP ingest server URL to enter in broadcasting software (e.g. OBS)
          type: string
        streamKey:
          description: Secret stream key — combined with rtmpUrl to start broadcasting. Never shown to anyone but the owner.
          type: string
        playbackId:
          description: Public Mux playback id for this channel's live stream
          type: string
...
    ChatMessage:
      type: object
      required: [id, channelId, userId, username, avatarUrl, message, createdAt]
      properties:
        id: { type: integer }
        channelId: { type: integer }
        userId: { type: string }
        username: { type: string }
        avatarUrl: { type: ["string", "null"] }
        message: { type: string }
        createdAt: { type: string, format: date-time }
    ChatMessageInput:
      type: object
      required: [message]
      properties:
        message: { type: string, minLength: 1, maxLength: 500 }
...